Refactor app shell and film detail UI components
